+++ endpoints/frontend.py
@@ -1,0 +1,20 @@
+from flask import request, Blueprint, Response
+import json
+import os
+
+front_blueprint = Blueprint('front_blueprint', __name__)
+
+@front_blueprint.route('/categoryBubbleData',methods = ['GET',"POST"])
+def get_categoy_bubble_data():
+   if request.method == 'GET':
+      #return json object that is used used to construct bubbles
+      print(os.listdir())
+      with open("topic_bubble_data.json") as f:
+         data = json.load(f)
+      print(type(data))
+      print("DATA:", data)
+      return Response(response= data)
+      #return data
+   elif request.method== "POST":
+      #update data in db - not implemented yet
+      return Response(response={},status=404)
